pcxt: enable full push_line editor and add missing i16 syscall stubs
Switch from push_line_stub.c to the real push_line.c so the shell gets
VT100 line editing, history, tab completion, and \w prompt expansion.

Add all 13 missing syscall stubs to ia16 syscall.S: ioctl, sigaction,
kill, nanosleep, ppoll, writev, readv, mkdir, unlink, rmdir, setsid,
setpgid, getpgid, ptrace. ia16 user-space now has full parity with
syscall.h declarations.

Fix pointer-to-long cast in push_line.c sigaction call for ia16 where
long is 32-bit but pointers are 16-bit.

+++ src/user/push_line.c
@@ -754,7 +754,7 @@ int push_readline(const char *prompt, char *buf, int size) {
 
   /* Install SIGINT handler */
   sigint_received = 0;
-  sigaction(SIGINT, (void *)(long)sigint_handler, 0);
+  sigaction(SIGINT, (void *)(uintptr_t)sigint_handler, 0);
 
   /* Enter raw mode */
   term_raw();
